Enable max warnings, fix library warnings, add /W4 to tests
CMakeLists changes:
- Remove -Wno-deprecated-declarations and -Wno-defaulted-function-deleted
  from GCC/Clang flags to surface all warnings
- Remove /wd4267 and /wd4996 from hmm target — deferred suppressions no
  longer needed after the refactor
- Add /external:anglebrackets /external:W0 to hmm target to reduce noise
  from MSVC STL headers
- Add /W4 /permissive- to test executables via add_hmm_test() so tests
  are held to the same warning standard as the library

Library warning fixes:
- C4101 (10×): remove unused 'e' variable from catch blocks across all
  10 distribution files — change 'catch (const std::exception& e)' to
  'catch (const std::exception&)'
- C4267 (1×): discrete_distribution.cpp:168 — add static_cast<int> when
  constructing DiscreteDistribution from size_t numSymbols
- C4996 (1×): file_io_manager.cpp — replace std::localtime with
  platform-safe localtime_s / localtime_r under #ifdef _WIN32
- C4244 (1×): file_io_manager.cpp — replace ::tolower with a lambda
  that casts via unsigned char to avoid int->char narrowing

Remaining 2 warnings: C4244 from MSVC's own <memory> AVX-512 template
internals. /external:W0 does not suppress them due to a known MSVC
limitation in template specialization warning paths. Accepted as
external-code noise; documented in CMakeLists comment.

36/36 tests pass.

+++ src/distributions/discrete_distribution.cpp
@@ -165,7 +165,7 @@ std::istream& operator>>(std::istream& is, libhmm::DiscreteDistribution& distrib
     
     // Create new distribution with the specified number of symbols
     try {
-        DiscreteDistribution newDist(numSymbols);
+        DiscreteDistribution newDist(static_cast<int>(numSymbols));
         
         // Read probabilities
         for (std::size_t i = 0; i < numSymbols; ++i) {
